requirements.txt
torch
transformers
pandas
scikit-learn
PyYAML
tqdm
jupyter
